refactor: layout text
